Clean up jQuery bindings for max compatibility and performance
The `jQuery( '#selector' )` pattern is known to be non-performant. A better alternative is to use native DOM methods for retriving an element by ID and pass that element itself into jQuery. This change has been made throughout.

Also, some older browsers have issues with `console.log` calls when the debugging console is not open. To avoid crashes and potential conflicts, we prefix this with a simple `'console' in window` test that, if false, will avoid running out debug log.